Added the images dataset

+++ CBIR/src/cbir/IndexedImage.java
@@ -6,5 +6,5 @@ import java.io.Serializable;
 public class IndexedImage implements Serializable {
 	private static final long serialVersionUID = 1L;
 	File file;
-	int[] imageDescriptor;
+	float[] imageDescriptor;
 }
